+++ src/Region/RegionHandler.h
@@ -92,6 +92,9 @@ public:
         CARTA::PvResponse& pv_response, GeneratedImage& pv_image);
     void StopPvCalc(int file_id);
 
+    // Image fitting
+    bool FitImage(const CARTA::FittingRequest& fitting_request, CARTA::FittingResponse& fitting_response, std::shared_ptr<Frame> frame);
+
 private:
     // Get unique region id (max id + 1)
     int GetNextRegionId();